<feat>初步调通了三节点三角形壳单元的代码逻辑,后续需要进一步修改算法逻辑中的bug

+++ src/component/property/PSHELL.h
@@ -15,7 +15,7 @@ class PSHELL : public PropertyBase {
  private:
   int _mid1;
   double _t;
-  double _mid2;
+  int _mid2;
   double _inertia_ratio_of_bm{1.0};
 };
 }  // namespace COMPONENT
